+++ lib/vagrant/provisioners.rb
@@ -2,7 +2,7 @@
 # as configuration classes right away with Vagrant.
 require 'vagrant/provisioners/base'
 require 'vagrant/provisioners/chef'
-require 'vagrant/provisioners/chef_server'
+require 'vagrant/provisioners/chef_client'
 require 'vagrant/provisioners/chef_solo'
 require 'vagrant/provisioners/puppet'
 require 'vagrant/provisioners/puppet_server'